Weave in CommenHtml in exports
You'll need to double check if this is the behaviour you wanted.
